+++ src/database/migrations/2018_01_03_210456_create_character_chat_channel_infos_table.php
@@ -16,10 +16,10 @@ class CreateCharacterChatChannelInfosTable extends Migration
 
         Schema::create('character_chat_channel_infos', function (Blueprint $table) {
 
-            $table->integer('channel_id')->primary();
+            $table->bigInteger('channel_id')->primary();
 
             $table->string('name');
-            $table->integer('owner_id');
+            $table->bigInteger('owner_id');
             $table->string('comparison_key');
             $table->boolean('has_password');
             $table->text('motd');
