+++ src/database/migrations/2018_01_05_053654_create_character_stats_table.php
@@ -16,7 +16,7 @@ class CreateCharacterStatsTable extends Migration
 
         Schema::create('character_stats', function (Blueprint $table) {
 
-            $table->integer('character_id');
+            $table->bigInteger('character_id');
             $table->integer('year');
             $table->string('category');
             $table->json('stats')->nullable();
