+++ src/database/migrations/2018_01_07_121524_create_character_contact_labels_table.php
@@ -16,7 +16,7 @@ class CreateCharacterContactLabelsTable extends Migration
 
         Schema::create('character_contact_labels', function (Blueprint $table) {
 
-            $table->integer('character_id');
+            $table->bigInteger('character_id');
             $table->bigInteger('label_id');
             $table->string('label_name');
